Refactor /send tests to their own file
